rename 'generate_term_cvs' to 'generate_final_csvs'
This is no longer just for term-*.csv
